% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_analyze_data.R
\name{analyze_outcome_glm}
\alias{analyze_outcome_glm}
\title{Analyze outcome using glm}
\usage{
analyze_outcome_glm(data, formula, family, data_aug)
}
\arguments{
\item{data}{data_frame which is assumed to have been trimmed already. It has to contain weight variables (\code{iptw1}, \code{iptw2}, \code{mw1}, \code{mw2}, \code{ow1}, and \code{ow2})}

\item{formula}{outcome analysis formula. It should only contain the outcome and the exposure of interest.}

\item{family}{\code{glm} family statement.}

\item{data_aug}{augmented data including counterfactual clones of each individual.}
}
\value{
data_frame with an adjustment method column and list columns for the coefficient and variance-covariance matrix (robust one for weighted data).
}
\description{
.. content for details ..
}
