% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_analyze_data.R
\name{augment_counterfactuals}
\alias{augment_counterfactuals}
\title{Create an augmented dataset of counterfactuals}
\usage{
augment_counterfactuals(data, outcome_name, counter_names, A_name,
  A_levels)
}
\arguments{
\item{data}{data_frame containing outcome, treatment, counterfactual outcomes under each one of three treatment values (0,1,2).}

\item{outcome_name}{name of the outcome variable.}

\item{counter_names}{names of the three counterfactual outcome variables}

\item{A_name}{name of the treatment variable}

\item{A_levels}{three levels for the treatment variable}
}
\value{
data_frame that is three times larger containing counterfactual outcomes for each individuals, weights are calculated from the true PS.
}
\description{
.. content for details ..
}
