% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{One-stop function for data preparation}
\usage{
prepare_data(data, formula1, formula2, family, ps_prefix1, ps_prefix2,
  df_trim_thres, A_name, levels)
}
\arguments{
\item{data}{data_frame}

\item{formula1}{PS model formula for the entire cohort estimation.}

\item{formula2}{PS model formula for the trimmed cohort estimation.}

\item{family}{PS model family}

\item{ps_prefix1}{PS variable prefix for the entire cohort estimation}

\item{ps_prefix2}{PS variable prefix for the trimmed cohort estimation}

\item{df_trim_thres}{data_frame with columns \code{trim_method_name} and \code{thres} to indicate}

\item{A_name}{Treatment indicator name in \code{data}.}

\item{levels}{Character vector holding three elements corresponding to levels in \code{A}}

\item{ps0_name}{Name of the column in \code{data} for the PS for the first level.}

\item{ps1_name}{Name of the column in \code{data} for the PS for the second level.}

\item{ps2_name}{Name of the column in \code{data} for the PS for the third level.}
}
\value{
nested data_frame containing
}
\description{
Takes a data_frame and
}
