% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_prepare_data.R
\name{trim_data}
\alias{trim_data}
\title{Dummy function to conduct no trimming at all.}
\usage{
trim_data(data, trim_method_name, thres, A_name, ps0_name, ps1_name,
  ps2_name, levels)
}
\arguments{
\item{data}{data_frame}

\item{trim_method_name}{Function used for trimming that takes \code{A}, \code{ps0}, \code{ps1}, \code{ps2}, \code{levels}, and \code{thres} as arguments. Either one of the string: \code{\link{none}}, \code{\link{crump}}, \code{\link{sturmer}}, \code{\link{walker}}.}

\item{thres}{Threshold for the trimming strategy.}

\item{A_name}{Treatment indicator name in \code{data}.}

\item{ps0_name}{Name of the column in \code{data} for the PS for the first level.}

\item{ps1_name}{Name of the column in \code{data} for the PS for the second level.}

\item{ps2_name}{Name of the column in \code{data} for the PS for the third level.}

\item{levels}{Character vector holding three elements corresponding to levels in \code{A}}
}
\value{
trimmed data_frame with fewer rows
}
\description{
.. content for details ..
}
