% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_prepare_data.R
\name{trim_walker}
\alias{trim_walker}
\title{Walker trimming function}
\usage{
trim_walker(A, ps0, ps1, ps2, levels, thres)
}
\arguments{
\item{A}{Treatment indicator vector having three levels indicated in \code{levels}.}

\item{ps0}{PS vector for the first level.}

\item{ps1}{PS vector for the second level.}

\item{ps2}{PS vector for the third level.}

\item{levels}{Character vector holding three elements corresponding to levels in \code{A}}

\item{thres}{Threshold for the trimming strategy.}
}
\value{
A numeric indicator vector. 1 for kept and 0 for dropped.
}
\description{
.. content for details ..
}
